+++ kqueue.go
@@ -181,6 +181,11 @@ func (w *Watcher) addWatch(name string, flags uint32) error {
 			return nil
 		}
 
+		// Don't watch named pipes.
+		if fi.Mode()&os.ModeNamedPipe == os.ModeNamedPipe {
+			return nil
+		}
+
 		// Follow Symlinks
 		// Unfortunately, Linux can add bogus symlinks to watch list without
 		// issue, and Windows can't do symlinks period (AFAIK). To  maintain